Major fixing of minor issues (lmao) v1.0

+++ SQL/schema.sql
@@ -30,10 +30,10 @@ is_circumbinary, orbital_period, orbit_radius, radius_earth, mass_earth, planet_
 CREATE TABLE Constellation (
     constellation_id INT AUTO_INCREMENT PRIMARY KEY,
     constellation_name VARCHAR(255) NOT NULL,
-    min_ra FLOAT NOT NULL CHECK (min_ra BETWEEN 0 and 360),
-    max_ra FLOAT NOT NULL CHECK (max_ra BETWEEN 0 and 360),
-    min_dec FLOAT NOT NULL CHECK (min_dec BETWEEN -90 and 90),
-    max_dec FLOAT NOT NULL CHECK (max_dec BETWEEN -90 and 90),
+    min_ra DOUBLE NOT NULL CHECK (min_ra BETWEEN 0 and 360),
+    max_ra DOUBLE NOT NULL CHECK (max_ra BETWEEN 0 and 360),
+    min_dec DOUBLE NOT NULL CHECK (min_dec BETWEEN -90 and 90),
+    max_dec DOUBLE NOT NULL CHECK (max_dec BETWEEN -90 and 90),
     description TEXT NOT NULL,
     pop_rank INT CHECK (pop_rank BETWEEN 1 AND 88) NOT NULL
 );
@@ -41,12 +41,12 @@ CREATE TABLE Constellation (
 CREATE TABLE StarSystem (
     system_id INT AUTO_INCREMENT PRIMARY KEY,
     system_name VARCHAR(255) UNIQUE NOT NULL,
-    `ra` FLOAT CHECK (`ra` BETWEEN 0 AND 360 OR `ra` IS NULL),
-    `dec` FLOAT CHECK (`dec` BETWEEN -90 AND 90 OR `dec` IS NULL),
+    `ra` DOUBLE CHECK (`ra` BETWEEN 0 AND 360 OR `ra` IS NULL),
+    `dec` DOUBLE CHECK (`dec` BETWEEN -90 AND 90 OR `dec` IS NULL),
     num_stars INT DEFAULT 1 CHECK (num_stars > 0),
     num_planets INT DEFAULT 1 CHECK (num_planets > 0),
     num_moons INT DEFAULT 0 CHECK (num_moons >= 0),
-    distance_pc FLOAT CHECK (distance_pc > 0 OR distance_pc is NULL),
+    distance_pc DOUBLE CHECK (distance_pc > 0 OR distance_pc is NULL),
     constellation_id INT,
     CONSTRAINT fk_constellation
         FOREIGN KEY (constellation_id)
@@ -74,13 +74,13 @@ CREATE TABLE Star (
         ON DELETE CASCADE,
     spectral_type VARCHAR(20),
     star_temp INT CHECK ((star_temp BETWEEN 415 AND 57000) OR star_temp IS NULL),
-    star_radius FLOAT CHECK ((star_radius BETWEEN 0.0115 AND 88.475) OR star_radius IS NULL),
-    star_mass FLOAT CHECK ((star_mass BETWEEN 0.0094 AND 10.94) OR star_mass IS NULL),
-    star_metallicity FLOAT CHECK((star_metallicity BETWEEN -1 AND 0.79) OR star_metallicity IS NULL),
-    star_luminosity FLOAT CHECK ((star_luminosity BETWEEN -6.09 AND 3.8) OR star_luminosity IS NULL),
-    star_gravity FLOAT CHECK((star_gravity BETWEEN 0.541 AND 8.07) OR star_gravity IS NULL),
-    star_age FLOAT CHECK ((star_age BETWEEN 0 AND 16.1) OR star_age IS NULL),
-    star_brightness FLOAT CHECK ((star_brightness BETWEEN 2.36431 AND 20.1861) OR star_brightness IS NULL)
+    star_radius DOUBLE CHECK ((star_radius BETWEEN 0.0115 AND 88.475) OR star_radius IS NULL),
+    star_mass DOUBLE CHECK ((star_mass BETWEEN 0.0094 AND 10.94) OR star_mass IS NULL),
+    star_metallicity DOUBLE CHECK((star_metallicity BETWEEN -1 AND 0.79) OR star_metallicity IS NULL),
+    star_luminosity DOUBLE CHECK ((star_luminosity BETWEEN -6.09 AND 3.8) OR star_luminosity IS NULL),
+    star_gravity DOUBLE CHECK((star_gravity BETWEEN 0.541 AND 8.07) OR star_gravity IS NULL),
+    star_age DOUBLE CHECK ((star_age BETWEEN 0 AND 16.1) OR star_age IS NULL),
+    star_brightness DOUBLE CHECK ((star_brightness BETWEEN 2.36431 AND 20.1861) OR star_brightness IS NULL)
 );
 
 CREATE TABLE Planet (
@@ -97,14 +97,13 @@ CREATE TABLE Planet (
         REFERENCES Discovery(discovery_id)
         ON DELETE SET NULL,
     is_circumbinary BOOLEAN NOT NULL,
-    orbital_period FLOAT CHECK ((orbital_period BETWEEN 0.090706293 AND 402000000) OR orbital_period IS NULL),
-    orbit_radius FLOAT CHECK ((orbit_radius BETWEEN 0.0044 AND 19000) OR orbit_radius IS NULL),
-    radius_earth FLOAT CHECK ((radius_earth BETWEEN 0.3098 AND 87.20586985) OR radius_earth IS NULL),
-    mass_earth FLOAT CHECK ((mass_earth BETWEEN 0.02 AND 9534.85221) OR mass_earth IS NULL),
-    planet_density FLOAT CHECK ((planet_density BETWEEN 0.0051 AND 2000) OR planet_density IS NULL),
-    eccentricity FLOAT CHECK ((eccentricity BETWEEN 0 AND 0.95) OR eccentricity IS NULL),
-    insolation_flux FLOAT CHECK ((insolation_flux BETWEEN 0.0003 AND 44900) OR insolation_flux IS NULL),
-    planet_temp FLOAT CHECK ((planet_temp BETWEEN 34 AND 4050) OR planet_temp IS NULL),
+    orbital_period DOUBLE CHECK ((orbital_period BETWEEN 0.090706293 AND 402000000) OR orbital_period IS NULL),
+    orbit_radius DOUBLE CHECK ((orbit_radius BETWEEN 0.0044 AND 19000) OR orbit_radius IS NULL),
+    radius_earth DOUBLE CHECK ((radius_earth BETWEEN 0.3098 AND 87.20586985) OR radius_earth IS NULL),
+    mass_earth DOUBLE CHECK ((mass_earth BETWEEN 0.02 AND 9534.85221) OR mass_earth IS NULL),
+    planet_density DOUBLE CHECK ((planet_density BETWEEN 0.0051 AND 2000) OR planet_density IS NULL),
+    eccentricity DOUBLE CHECK ((eccentricity BETWEEN 0 AND 0.95) OR eccentricity IS NULL),
+    insolation_flux DOUBLE CHECK ((insolation_flux BETWEEN 0.0003 AND 44900) OR insolation_flux IS NULL),
+    planet_temp DOUBLE CHECK ((planet_temp BETWEEN 34 AND 4050) OR planet_temp IS NULL),
     ttv_obs BOOLEAN NOT NULL
 );
-
